partially done with game

+++ pages/api/remove.js
@@ -12,21 +12,25 @@ export default async function handler(req, res) {
   const client = await clientPromise;
   const db = client.db("bsblDB");
   let isAuthenticated = false;
-  // check valid authentication, if username and passwords count. then check if id is in queue right now
-  await db
-    .collection("account")
-    .findOne({ email: req.body.email })
-    .then((u) => {
-      if (u) {
-        if (u.password == req.body.password) {
-          // authenticated
-          isAuthenticated = true;
+  if (req.method == "POST") {
+    // check valid authentication, if username and passwords count. then check if id is in queue right now
+    await db
+      .collection("account")
+      .findOne({ email: req.body.email })
+      .then((u) => {
+        if (u) {
+          if (u.password == req.body.password) {
+            // authenticated
+            isAuthenticated = true;
+          }
         }
-      }
-    });
-  if (isAuthenticated) {
-    await db.collection("queue").deleteOne({ uid: req.body.uid });
-    res.json({ success: true });
+      });
+    if (isAuthenticated) {
+      await db.collection("queue").deleteOne({ uid: req.body.uid });
+      res.json({ success: true });
+    }
+    res.json({ success: false });
+  } else {
+    res.json({ success: false });
   }
-  res.json({ success: false });
 }
